Add support for GPUs without instanced rendering
The patch 31b3088 introduced instanced rendering for efficiently
updating only the damaged area of the output. However due to this new
feature, support for older GPUs without instanced rendering was dropped.

This patch adds an alternative implementation which is chosen at runtime
when the target GPU does not support instanced rendering. It uses a
combination of uploading duplicate information and multiple draw calls
to somewhat efficiently mimic the effects of instanced rendering.

I've verified that with this patch it is now possible to run Anvil on my
PinePhone.