x11: use EGLSurface instead of dmabufs via dri3
Systems without dri3 are still widely in use
(mostly with regards to the proprietary nvidia driver).
Given these unfortunate circumstances our X11Backend
should not depend on DRI3 and Present to be available.

This commit changes the X11 backend code to utilize
an EGLSurface instead, utimately working very similar
to the winit backend.