Remove winit constraint from EGLNativeSurface impl
To make it possible for wayland clients to use Smithay's EGL
abstractions, it's necessary to have an available implementation for
`EGLNativeSurface`.

Currently the implementation for `WlEglSurface` is only available with
the `backend_winit` feature enabled, which pulls in a ton of additional
dependencies. Since it's not actually required, just implementing it
with the `backend_egl` enabled should be more appropriate.