aggregate the surface offset on render
the surface offset is used to move the provided
damage (which is relative to the passed location)
to the relative position of the surface in the tree
in order to calculate the surface local damage
intersection.
if we do not agggregate the surface offset the
relative damage location will be wrong for
subsurface trees with a depth greater than 1.