input: Add a way to add a keyboard to the seat from an keymap string
This should provide a way to fix things like
https://github.com/pop-os/xdg-shell-wrapper/issues/8.

Cosmic's panel is implemented as a compositor that hosts applets as
clients. It should just pass through the same keymap that is provided by
the host compositor, rather than trying to create one from any
configuration.

This could also be useful for the Wayland backend
(https://github.com/Smithay/smithay/pull/676), depending on use case.
Then a compositor using the Wayland backend would respect the host
compositors keymap settings, but for some testing it may be helpful to
do otherwise.

This cannot directly accept an `xkb::Keymap` due to how `Keymap` uses
reference counting and is not thread safe. Or rather it could, but only
by converting it to a string and back internally.