Add pointer constraints protocol
A pointer constraint is unique per surface and pointer. Creating a
constraint when one already exists is a protocol error. Technically the
spec says it should be a protocol error to add a constraint for the same
"seat", but as things are implemented in Smithay it's easier to tie it
to the `PointerHandle`.

This provides `activated`/`deactivated` helpers, which is somewhat
similar to the API wlroots provides.

Adds implementation in anvil. This implementation doesn't currently
handle regions, which requires a non-trivial algorithm / data structure.
But that should be good enough for an initial implementation in Anvil.