renderer: report the render state of the elements
this extends the result of the dtr with a report
that holds the render states of the provided elements.
The state can be used to query states like skipped after
the rendering took place. This enables use-cases
where the actual state is needed like frame
callback throttling and wp_presentation.