Set MSRV in Cargo manifest
This sets the minimum supported Rust version to 1.65.0 in Smithay's
manifest, which helps with troubleshooting downstream when attempting to
build Smithay with an outdated compiler.

Especially since the MSRV is far above the introduction of the
`rust-version` field anyway, it should not come with any significant
drawbacks besides the potential maintenance overhead.