Fix IME not closing on application exit
This fixes an issue with the text_input/input_method protocols where the
application requesting IME wouldn't notify the IME client about its
death, keeping the IME client open after the application was closed.

With this patch whenever the `text_input` object is destroyed, the
`deactivate` event is automatically emitted and committed ensuring the
IME client is properly updated.

Closes #835.