Use pointer and length to refer to shm memory, instead of slice
Creating a slice into shared memory is problematic, since this is
considered undefined behavior if the client mutates the memory. This is
especially a problem since it was creating a slice for the whole pool,
as even a correctly behaving client may mutate another buffer in the
same pool.

We don't want to trust clients not to mutate buffers for soundness, even
if it's not obvious what bugs or exploits this could result in. It's
also quite unnecessary if all we do is convert the slice back to a
pointer and pass it to OpenGL.

Memory access that could result in `SIGBUS` is probably also problematic
in access from safe Rust code that expects all memory accesses to
succeed.

This simply replaces the slice with a pointer and length, which for the
most part (with the current renderer implementation) means we're not
converting to a slice just to convert it back to a pointer again.