gbm: export fd per plane
the recent IO safety work revealed a bug
in GbmBuffer::export when the system does
not support gbm_bo_get_fd_for_plane.
Before the IO safety work the same fd was
used for multiple planes (this is problematic
because each plane takes ownership of the fd).
Practically this should not have happened as
the legacy version only supports bo with a single
plane.

The latests gbm release made this kind of
error impossible by directly returning an OwnedFd,
but we missed it in smithay because of conditional
compilation.