backend: Gesture events
This provides an implementation for libinput. To be useful in
compositors, an implementation of the `zwp_pointer_gestures_v1` Wayland
protocol is also needed.

Hold gestures require `libinput_1_19` so the code for that in the
libinput backend is behind a feature flag.

(Would it be better to implement this without 10 new traits and 8
associated types? In any case, this is consistent with how other event
types are handled.)