Fix render output transforms
The transform of an output is specified in the rotation that needs to be
applied to a surface for rendering. This is the exact inverse of the
transform that needs to be applied to the renderer itself.

This patch inverts the transform stored on the output's mode to ensure
the rendering transform correctly specifies the inverse of the surface
transform.