wayland.xdg-decor: Send `configure` if it has not been sent yet
Previously this only sent configure if it had changed, but this doesn't
work when a client doesn't set a preferred mode and relies on the
server to define the mode.

This adds an additional variable for whether the decoration configure
event has been sent, since I don't see any indication that
`get_toplevel_decoration` couldn't be called some time after the first
`configure`.

The protocol spec also states it is a protocol error to create a
`xdg_toplevel_decoration` with a buffer attached, or to attach a buffer
before the `xdg_toplevel_decoration.configure` is received. It may be
good to enforce that, though perhaps first check the behavior of clients
and other compositors. The fact `winit` windows ever showed anything
before this change suggests it isn't following this restriction...

Fixes decoration of winit windows in cosmic-comp (and presumably other
things).