xwm: Fix setting `_NET_ACTIVE_WINDOW`
This is probably not the best way to handle this, but I'm not sure what
is. It requires a new root argument to `X11Surface::new`, which is a
change to the public API though this function is mainly called within
Smithay itself, it seems.

Previously this was set based on the `FocusIn`/`FocusOut` events of the
root window, but those always have the id of the root window. Presumably
X sends the events to the focused window and all parent windows. So
applications saw a change to `_NET_ACTIVE_WINDOW`, but the value was
always the id of the root window. This seems to behave more correctly.