xwm: Only set `_NET_WM_STATE` if it has changed
EWMH says a client MUST use a client message to change this, so this
should be managed by the compositor and it is not necessary to read the
value from X. Assuming the `X11Surface` exists for the lifetime of the X
window, and the compositor only sets this via `change_net_state`.

Fixes https://github.com/pop-os/cosmic-comp/issues/90. This should also
generally perform better since it removes a `get_property` call which
has to wait on the X server, and it doesn't allocate a `HashSet`. For
what difference that makes.