Fix virtual keyboard crash with invalid keymap
Since the keymap sent for the virtual keyboard protocol is supplied by a
client, it should not be trusted. Unwrapping on it causes the compositor
to crash when a malicious or buggy client tries to send the `Keymap`
request.

To ensure even misbehaving clients don't cause any issues, requests with
invalid keymaps are simply ignored.

This has been sparked by an issue encountered with `wvkbd`, which sends
a keymap with 54 `\0` bytes at the end. While it would be possible for
Smithay to clean this up, I'm not sure if it is appropriate for the
compositor to work around what is arguably a client issue.

An issue has been filed with `wvkbd` here:
https://github.com/jjsullivan5196/wvkbd/issues/35

Closes #942.