Rename FractionScaleHandler to FractionalScaleHandler
Both the protocol and the existing `FractionalScaleManagerState` name
this feature "fracitonal scale". During implementation I just quickly
typed out the code and had to fix it after realizing that Smithay's
handler has inconsistent naming.

While this is a somewhat unnecessary breakage, I think in the long run
it's going to be beneficial to have consistent naming that doesn't get
in the way and this feature is still relatively new.