Truncate nul bytes from virtual keyboard string
This fixes a regression introduced in e16cee1, which caused the virtual
keyboard protocol to break completely with a `NulError` in xkbcommon.

Since xkbcommon uses `CString::new`, which panics whenever encountering
a nul byte, it is important to first strip any nul bytes from our
String. Without this, the function is guaranteed to panic since there
will always be at least on trailing nul byte.