drm: fix element switching direct scan-out <> composition
When en element switches between direct scan-out and
composition we will dynamically add or remove fake
elements to communicate the opaque regions of overlay
planes and realize hole punching in case of underlays.
Re-using the id of the elements will miss with the damage
state of the internal output damage tracker. Instead we now
use an artifical Id per plane/element for the fake elements.
This now also reports the original element damage in the
OverlayPlaneElement, this makes sure that damage for partially
opaque elements hit the primary plane.