SubsurfaceUserData: Add an accessor for the surface
Wayland defines subsurfaces in terms of a surface with a role. Smithay
also treats it as such and this is problematic for protocol that is
designed to work with wl_subsurface objects directly. Without a
mechanism to obtain the underlying protocol implementations external to
Smithay are unable to perform any operations.

Here is an example of usage using Chrome specific protocol (whose
actually usefulness is out of scope).

augmented_sub_surface::Request::SetPosition { x, y } => {
    if x.trunc() != x || y.trunc() != y {
        info!("Unsupported float offset x:{x} y:{y}");
    }
    with_states(
        data.wl_subsurface
            .upgrade()
            .unwrap()
            .data::<SubsurfaceUserData>()
            .unwrap()
            .surface(),
        |states| {
            states
                .cached_state
                .pending::<SubsurfaceCachedState>()
                .location = (x as i32, y as i32).into()
        },
    );
}