input: Add `KeyboardHandle::set_xkb_config` method
This allows changing the keymap used the keyboard without removing
and re-adding the `wl_keyboard`. This seems to be the best way to do
dynamic xkb configuration, so settings can be changed without restarting
the compositor.

Like `wlr_keyboard_set_keymap`, this creates a new `xkb_state`, updates
it with pressed keys, and sends modifiers to clients.