Remove `Boottime`
This doesn't seem to exist on platforms like NetBSD. Getting system
uptime is possible, but involves `sysctl` and not a `clockid_t`.

OpenBSD seems to have both `CLOCK_BOOTTIME` and `CLOCK_UPTIME`. But they
are different (`CLOCK_UPTIME` excludes time spent suspended.)

This isn't used anywhere. So unless it's expected to be important for
something, may be simplest to remove it?