Use a thread-local for thread ID in `UserDataMap::get`
Profiling suggests the compositor was spending a surprising amount of
time getting the thread ID. It seems `thread::current().id()` is a bit
slow. Calling this once and storing the result in a thread-local seems
to make the performance roughly match the thread-safe variant, if the
benchmark added here is accurate.