Reduce log level of common warnings
Log levels of warning and above should indicate unexpected behavior to
the user. If the application is performing as expected, users shouldn't
be misguided by log messages that are not relevant to potential issues.

If an issue is encountered because of these warnings without any other
relevant details in the log, then the log level can still be increased
to get a full debug log.