Add buffer age reset function
The PPP's driver seems to run into non-recoverable framebuffer errors
when the `reset_buffers` function is called and the `InternalSlot` is
reset while the `InternalSlot::buffer` is kept.

This patch adds a new function which only resets each slot's buffer age
rather than resetting the entire slot, allowing a force-redraw on the
PPP without triggering the framebuffer error.