Remove offset from seat's touch API
The offset passed along with the location to calculate the relative
position of the touch event was an idiosyncrasy introduced due to the
way I structured the data in my compositor in the past.

At this point even my compositor doesn't use this structure anymore and
just stores a relative position, so calculating the offset back from
that is a waste of effort and prone to errors.

So to simplify the API, the touch down event now just takes the
surface-local position, identical to the way the protocol itself
behaves. Documentation has been added to clarify this.