Expose swapchain's reset_buffer_age function
In a1e4cf3, the new `reset_buffer_age` function was added to the
swapchain for resetting damage without having to reset the buffers
themselves.

This function is not used anywhere in Smithay but could be a simpler and
more clear API for users to "force" redraws. However the swapchain
itself is not directly accessible to users.

This patch adds a public `reset_buffer_ages` function to all structs
providing a `reset_buffers` method as an alternative for Smithay's
consumers.

The original `reset_buffer_age` function has also been renamed to
`reset_buffer_ages` to be more in line with `reset_buffers` and signify
that there's multiple buffers being involved here.