Allow customizing GlesRenderer capabilities
This patch separates the retrieval of the available driver capabilities
from `GlesRenderer::new`, allowing users to define their own capability
list.

The driver's defaults can be easily retrieved with the new
`Capability::supported_capabilities` function.

The `None` value is allowed to mimic the current behavior to ensure the
complexity of the API isn't increased unnecessarily for people not
interested in this feature.

Closes #1023.