WIP Use `PopupKind`/`PopupManager` for input method surface
Fix for https://github.com/Smithay/smithay/issues/1045. This eliminates
the need for `set_point`, and for the compositor to have separate code
for rendering IME popups.

This seems to work at least as well as the previous version with fcitx5
in anvil? The use of `set_point` previously wasn't correct. But some
work is still needed.

The current version of the protocol also has some fundamental issues
that may be problematic here. We can bring anything like that up in
discussion of the v3 protocol.