egl: Only query eglGetError() when an error happened
Most EGL calls return a value which indicates whether the call succeeded or
not, so avoid checking for errors when none happened.

I went for specialisation since checking for != EGL_FALSE gets redundant
pretty quickly, same for !.is_null().  There are a few functions
returning unexpected errors though, in our case eglClientWaitSync() and
eglDupNativeFenceFDANDROID(), returning respectively EGL_FALSE even
though it is a EGLint, and -1.

This has been tested with the Catacomb[1] compositor, it indeed reduces
the amount of eglGetError() calls to exactly zero according to apitrace.

[1] https://github.com/catacombing/catacomb