Add back manual `Debug` impls removed in f628336b
These are still needed to implemented `Debug` without a `D: Debug`
bound.

If there were *one* manual debug impl like this I'd add a comment noting
that it is the same as the derived implementation but with a different
bound, and link to https://github.com/rust-lang/rust/issues/26925. But
adding that comment everywhere would be verbose...

This still slightly simplifies many of these implementations, since they
used to require `Debug` bounds for `::KeyboardFocus`, `::PointerFocus`
etc. When removing those I switched to deriving `Debug` since that
*seemed* to work.