egl: Support ANGLE EGL platform on winit Wayland
For this to work I also had to modify `LIB` to point to the `libEGL.so`
from ANGLE. And set `VK_ICD_FILENAMES`, because naturally it wasn't
working with Nvidia drivers, but was with Intel.

This uses Vulkan. It seems ANGLE only provides the Wayland platform with
Vulkan.

This then runs, but errors with "Depth/stencil buffer format combination
not allowed for blit." It seems it needs to use `DEPTH24_STENCIL8`.
Presumably that's what the default framebuffer is using?