anvil: Don't store `pointer_location`, use from `PointerHandle`
If the pointer location is tracked within Smithay, it seems bad for to
have a separate copy of it; assuming we never want them to get out of
sync. And it seems in a couple cases `pointer_location` was being
changed without emitting pointer events and changing the value within
Smithay.

The `PointerHandle` is stored in state, so it isn't necessary to call
`get_pointer` and `unwrap` everywhere. Which seems good to avoid anyway.

In the few cases where `pointer_location` was changed without emitting
pointer motion events (like with tablets) this now emits pointer events
as well.