WIP Use euclid crate for geometry types
This does not yet compile.

It seems one difference between the Euclid types and the ones Smithay
currently has is that `euclid`'s points cannot be added, and there's a
seperate vector type. That is probably better for making sure code
dealing with coordinates is correct. Though that will require some
changes.