Ei protocol support using `reis`
The Ei protocol is needed for the xdg-desktop-portal `RemoteDesktop`
portal to emulate input devices, as well as by the `InputCapture` portal
for Synergy-like uses (input-leap supports Wayland with this portal).

This is quite incomplete, but the `type-text` example in `reis` now
works in Anvil. Reis could still use somewhat better higher-level
servere-side APIs.

It might make sense if ei exposed seats and devices matching those in
Smithay (or would compositors not always want to do that?). Not sure how
best to handle that.

Receiver contexts for the `InputCapture` portal are also a bit more
complicated to implement. Those involve capturing input once the cursor
crosses outside the display. That isn't implemented at all here yet.