Make XdgShellHandler::reposition_request mandatory
When running a GTK4 application which tries to reposition a popup
surface, it will wait for a commit to confirm the repositioning and halt
all rendering.

To avoid compositors accidentally skipping the implementation of this
function, it should be clearly indicated that this is mandatory. The
documentation was also expanded to reinforce this point.