Start global serial counter at 1
Qt apps don't acknowledge the xdg_surface configure event when the
serial number is 0, therefore, no toplevel appears.

This bug never happens on most compositors because the serial number 0
is used for other things, and by the time a configure event comes, the
number is already incremented.

The minimal example suffers from this bug because it's so minimal, and
unfortunately serial number 0 is used for the first configure event.