Add separate keymap for virtual keyboard
Smithay's virtual keyboard implementation previously stored a single
keymap on the seat, updating it whenever either the real or virtual
keyboard's keymap was changed. As a result the two keyboards could break
each other by setting an incorrect keymap.

To allow both the seat's real keyboard and the virtual keyboard to
function simultaneously, a separate keymap is stored for each and
whenever an input event related to a keymap is handled, the keymap is
updated.

Closes #1222.