High res scrolling support
This adds support for `wl_pointer` version 8, calling
`wl_pointer::axis_value120` with the v120 values from libinput. For
clients not binding version 8, values are accumulated to multiples of
120 to send the old `axis_discrete` event.

This helps even for clients using `wl_pointer::axis` with old versions
of `wl_pointer`, since the old libinput event only was emitted where
there was a whole discrete scroll step.

Raises libinput requirement to 1.19, which should be widely available
now. Removes handling of versions before that.

It should be possible to support high res scrolling in the X11 and Winit
backends, where the platform supports it, but for now this just converts
the low-res value to v120.