popups: Emit a wl_pointer.leave when destroyed
Wayland protocol indicates that a leave event must be generated before
any subsequent enter event. WLCS (but true in general) have several
tests where a popup is destroyed and it is expected to leave and
re-enter the below surface.

Half of the issue is addressed here, when a popup is explicitly
destroyed the leave event will be generated, though the subsequent enter
will not be generated until there is pointer motion. The enter event can
be simulated within the compositor via a 0 length move event, but the
leave event has no such workaround.

This issue is potentially present for surfaces and subsurfaces as well,
but WLCS doesn't seem to care (that I've yet seen) about those - so
that's deferred until needed.

Fixes #1257