dmabuf: Store single modifier, instead of storing per-plane
Version 5 of the protocol requires an error if planes do not all have
the same modifier:
https://gitlab.freedesktop.org/wayland/wayland-protocols/-/merge_requests/224.

Separate modifiers per-plane is apparently a historical artifact that
isn't actually supported in drm in the kernel. So it makes sense to
remove it from the `Dmabuf` abstraction in Smithay. The `Buffer`
implementation already used just the modifier for the first plane.

This has the logic to produce an error when version 5 of the protocol is
exposed, but still creates the global with version 4 until
wayland-protocols is updated.