drm: do not silently drop gbm bo usage flags
in case we miss support fo create_buffer_object_with_modifiers2
we can only fall back to create_buffer_object_with_modifiers if
we have a certain subset of usage flags.
if we got flags we can fall back to create_buffer_object for invalid
and linear modifiers, but not for others.

fixes cursor usage flags which is required for the bo on the cursor plane.