drm: fix inactive device error race
it can happen that a device is already
inactive while we try to use it. the
code already handled that by testing
for various error codes, but it seems
that check broke at some time.
`raw_os_error` only returns the
error code if it was initialized by  last_os_error or from_raw_os_error.
Otherwise None is returned and we
fail to detect the error.