desktop: check if popup is still alive
when searching for a popup of a surface we
have to test if the popup we found is still alive
or otherwise we might return a destroyed popup.
normally this wouldn't be an issue and the popup
should get cleaned up eventually, but it can race
against cleanup when a surface is re-used. re-using
the surface requires that the surface has no buffer
attached on the first commit, so it can happen that
directly after the popup is destroyed a NULL buffer
is committed. In this case we could end in a situation where we try to send the initial configure to the dead popup and afterwards skip sending the configure to the re-used popup.

fixed deadlock in epiphany (Gnome Web) when
selecting an history entry in the title bar.