input: allow PointerTarget to handle focus replace
this adds a new function to PointerTarget that
is called when the focus is replaced. this allows
implementing a nested stack of focus targets
without sacrificing grab guarantees.

the default impl will do the same like the old
behavior, so call leave on the previous target
and enter on the new one.