desktop/window: Fix `bbox()` for XWayland Windows
This fixes a bug with the latest version of Epiphany not rendering/working,
when run in XWayland, on anvil or cosmic-comp, as reported on
https://github.com/Smithay/smithay/pull/1319. Presumably it fixes
similar issues reported with other client; not sure why this doesn't
break XWayland more consistently...

`xtrace` shows `ConfigureNotify` was being sent with a size of `(0, 0)`
because of this. If `Window::bbox` calls `X11Surface::bbox`, the
behavior should be the same as it was before the `Window` change.

This probably isn't the best way to do things, with a `bbox` field
that's only used for Wayland toplevels. Not sure what would be better.