Run Xwayland using -displayfd and a pipe
This method, while undocumented, is a lot more robust than an intermediate
shell. It works by passing a pipe to the Xwayland cmdline invocation. When
Xwayland is ready, it writes the display number and a newline to the pipe.

While we're at it, we can remove the hard dependency on Calloop, so that the
Xwayland instance can be used with other event loops.