Add `unset` method to input grab traits
`unset` is called when a grab is removed with `unset_grab` or replaced
with `set_grab`.

This fixes https://github.com/pop-os/cosmic-comp/issues/403. There are a
couple other grabs in Smithay and cosmic-comp that they run before
unsetting the grab, that may cause issues if the grab is removed without
running it. So this seems useful in general.