Allow `set_cursor` when there is a pointer grab for client
This also updates `cursor_shape` call the same logic.

This fixes calls to `set_cursor` during drag-and-drop. Nautilus is a
good client to test this with, though most drag-and-drop clients should
work.

The Wayland spec doesn't seem that specific about how `set_cursor`,
pointer focus, and drag-and-drop interact. But this seems to mostly
match Sway: the surface loses pointer focus at the start of
drag-and-drop, but can still set the cursor.

I'm not sure if this is likely to be desirable or undesirable when there
is a grab other than a drag-and-drop grab. It could be restricted to
drag-and-drop, but there is not currently any way in Smithay to check
what the drag is, or if there is a drag-and-drop operation. But it
should be possible to downcast the `dyn PointerGrab<D>`, I guess.