wlr_layer: Fix memory leak due to reference cycle
`WlrLayerSurfaceUserData` contains a `WlSurface`.
`LayerSurfaceAttributes`, which is stored in the surfaces `data_map`,
contains a `ZwlrLayerSurfaceV1`. This, somewhat subtly, is a reference
cycle that causes a leak. So one of these references needs to be weak.

I noticed this issue with Valgrind, but this can also be reproduced by
adding  drop impl to `LayerSurfaceAttributes` that panics. Before this
change, running `simple_layer` in anvil and exiting it never drops the
`LayerSurfaceAttributes`. With the change, the panic happens when a
layer surface is closed.