Call `wl_buffer::release` only in drop impl of `renderer::utils::Buffer`
The protocol says a buffer that is attached but then replaced without
being committed should not get `release`. So there shouldn't be any
reason to call `release` in `merge_into` or `cleanup`.

Once the surface is comitted, `RenderSurfaceState::update_buffer` will
create a `Buffer`. So it seems everything should be correct to ensure a
buffer is released once and only once, when there are no references.