Check if focus is unchanged in `set_data_device_focus`/`set_primary_focus`
I wanted to change where these are called in `cosmic-comp` to workaround
an issue. But by not just calling it when keyboard focus changes, it
created a flood of `offer` events.

It would be possible to add a getter for the focus, but it seems
reasonable just to add a test here to check if it has changed. I assume
there's no problem doing noting if the focus is the same as the current
one.