damage: simplify opaque region handling
this does a few things, first it removes some redundant rectangle subtractions.
it is sufficient to subtract it once from the damage at the end instead of multiple
times on the intermediate results. this might also result in less fragmented damage
for the shaper.

it also gets rid of the vec in vec and instead of doing implicit indexing using explicit
indexing in a (linear) single vec. this makes access a lot cheaper, especially for
subtracting opaque regions above the currently drawn element.