Refactor fractional-scale to store state persistently
In my compositor I always set scale for every surface as it changes
outputs and such. It is therefore convenient if the fractional scale is
always stored in Smithay and automatically sent to newly-bound
fractional scale objects. This way, I don't need to recompute it
whenever a new fractional scale object is created.