drm: always submit the complete state
while it should be totally fine to only
update the planes that have actually changed
it seems some drivers might break by doing so.

after some trial and error it seems that it might
be enough to always include the primary plane
in the update to workaround this issues. But
to be safe for now and until this has resolved
we just include all used planes in the update.